Add inline-block to react-datepicker__input-container
`input` elements are inline-block but `div` is a block element.

react-datepicker wraps its input with a div. As far as I can tell, the
purpose of the wrapper is setting up a positioned container for the
absolutely positioned `react-datepicker__close-icon`. But since the
`input` isn't a block element, the close-icon doesn't line up with the
`input`. Making the wrapper `div` have `display: inline-block` fixes
that.