Fix positioning of calendar when month is changed
For some months there is an extra row of days. Tether component calls
position method of tether instance but at the moment that method called
react doesn't update the real dom. As well as fixing this asynchronous
problem, there is an another bug in tether library, which can not
recalculate the position for the first call of `position()` method. I
observed if you call that method many times (between 1 - 10) it finally
recalculate correctly.