ShortMonth on MonthDropdown
For this to work I added prop useShortMonthInDropdown in DatePicker & MonthDropdown components
Also added the function getMonthShortInLocale in date_utils and the conditional in Calendar component

Finally added a className for selected option on both Month & Year Dropdown selected option. This only for css customization, because can't format parent based on child options.