Added public function for programtically clearing the input
I've added a clear function intended to be publically accessible. It actually just runs the onClearClick function of `DatePicker`, but doesn't pass an event object to it as it may not be available. In order to make the onClearClick function not crash I've also added a quick check to make sure an even object was actually passed to the function before attempting to run `event.preventDefault()`.