Calculate scroll of TimeSelect
Changed hardcoded 30px-per-item scroll in order to have selectedDate/currH scrolled down in the middle of the TimeSelect. Now it saves a reference of list-item of selectedDate (with fallback to currH when selectedDate not provided) and calculates list's default scroll based on list's height and the referenced item's offset.

Add tests