Update to react 16 createRef API
I am not sure how customInputRef is meant to work as any value but ref seems to result in a console warning.

What I want to do is pass in my own reference to the input so I can have access to the dom node to check its size. The createRef API works great but this has not been updated and overwrites the ref I put on the custom input. I can work around this by passing a dummy value to customInputRef but that seems kind of silly and puts a warning in the console.