Don’t display time as selected when selected prop is not present
The previous behavior was very strange; it would display the current
time as selected, but only if the current time exactly matched one
of the time selections. E.g if you open the time window at 4:00 you will
see that time as selected, but not if you open it at 4:01 (assuming
greater than 1 minute interval).