Open to time (#1748)
* Scroll to time based on openToDate prop

When openToDate prop is present and selected prop is not, set the active
(visible) time based on openToDate rather than the current time.

This addresses https://github.com/Hacker0x01/react-datepicker/issues/1150.

* Don’t display time as selected when selected prop is not present

The previous behavior was very strange; it would display the current
time as selected, but only if the current time exactly matched one
of the time selections. E.g if you open the time window at 4:00 you will
see that time as selected, but not if you open it at 4:01 (assuming
greater than 1 minute interval).

* Correctly calculated the closest time to the nearest interval

Previously it was setting active time to the top of the hour, unless there
was an exact match.