Year change day selection
When a year change was initiated via the Year `select` dropdown, the date would not be focused in the calendar. This meant that keyboard users would be stranded in the calendar header, unable to tab to the date.

This change uses Dan Diaz’s work on `handleMonthChange` (#2097) to give the same functionality to `handleYearChange` and restore keyboard accessibility.