fix: attempt to validate dates when minTime and maxTime props are specified
I'm trying to address #2720 (https://github.com/Hacker0x01/react-datepicker/issues/2720). The value of the input should be reverted to date AND time that is within allowed constraints. This already happens for `maxDate` and `minDate` properties so it should be consistent for times as well.

However I am not able to pinpoint in the code where this normalization
occurs. When user presses enter or blurs the input, the value should
revert. I've tried adding validation code to parts where I thought this
occured but I am having no such luck.