fix: Ensure accurate scroll position on time component mounts
This commit addresses an issue where the scroll position was not accurately set when the component mounted.  To resolve this, we have incorporated the use of requestAnimationFrame to delay the scroll update until after the component has fully rendered and layout calculations are complete.  This ensures consistent and precise scroll behavior.

Closes #4290