fix: Update home key and end key navigation in Calendar component
Adjust home and end handlers to align with W3C WAI-ARIA principles.  Previously navigation moves to the first and last day of the week instead of one year before and after the calendar.  After this change, it's moving to [the first and last day of the week](https://www.w3.org/WAI/ARIA/apg/patterns/dialog-modal/examples/datepicker-dialog/)

Closes #4420